KTB-15 created standalone class with telegram logic and parsing
